third_party/blink/web_tests/external/wpt/css/css-typed-om/idlharness-expected.txt

This is a testharness.js-based test.
Found 59 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] CSSNumericValue interface: calling parse(USVString) on unitValue with too few arguments must throw TypeError
  assert_own_property: interface object must have static operation as own property expected property "parse" missing
[FAIL] CSSStyleValue interface: calling parse(USVString, USVString) on unitValue with too few arguments must throw TypeError
  assert_own_property: interface object must have static operation as own property expected property "parse" missing
[FAIL] CSSStyleValue interface: calling parseAll(USVString, USVString) on unitValue with too few arguments must throw TypeError
  assert_own_property: interface object must have static operation as own property expected property "parseAll" missing
[FAIL] CSSNumericValue interface: calling parse(USVString) on mathSum with too few arguments must throw TypeError
  assert_own_property: interface object must have static operation as own property expected property "parse" missing
[FAIL] CSSStyleValue interface: calling parse(USVString, USVString) on mathSum with too few arguments must throw TypeError
  assert_own_property: interface object must have static operation as own property expected property "parse" missing
[FAIL] CSSStyleValue interface: calling parseAll(USVString, USVString) on mathSum with too few arguments must throw TypeError
  assert_own_property: interface object must have static operation as own property expected property "parseAll" missing
[FAIL] CSSStyleValue interface: calling parse(USVString, USVString) on transformValue with too few arguments must throw TypeError
  assert_own_property: interface object must have static operation as own property expected property "parse" missing
[FAIL] CSSStyleValue interface: calling parseAll(USVString, USVString) on transformValue with too few arguments must throw TypeError
  assert_own_property: interface object must have static operation as own property expected property "parseAll" missing
[FAIL] CSSColorValue interface: existence and properties of interface object
  assert_equals: prototype of CSSColorValue is not CSSStyleValue expected function "function CSSStyleValue() { [native code] }" but got function "function () { [native code] }"
[FAIL] CSSColorValue interface: existence and properties of interface prototype object
  assert_equals: prototype of CSSColorValue.prototype is not CSSStyleValue.prototype expected [stringifying object threw TypeError: Illegal invocation with type object] but got object "[object Object]"
[FAIL] CSSLab interface: existence and properties of interface object
  assert_own_property: self does not have own property "CSSLab" expected property "CSSLab" missing
[FAIL] CSSLab interface object length
  assert_own_property: self does not have own property "CSSLab" expected property "CSSLab" missing
[FAIL] CSSLab interface object name
  assert_own_property: self does not have own property "CSSLab" expected property "CSSLab" missing
[FAIL] CSSLab interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "CSSLab" expected property "CSSLab" missing
[FAIL] CSSLab interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "CSSLab" expected property "CSSLab" missing
[FAIL] CSSLab interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "CSSLab" expected property "CSSLab" missing
[FAIL] CSSLab interface: attribute l
  assert_own_property: self does not have own property "CSSLab" expected property "CSSLab" missing
[FAIL] CSSLab interface: attribute a
  assert_own_property: self does not have own property "CSSLab" expected property "CSSLab" missing
[FAIL] CSSLab interface: attribute b
  assert_own_property: self does not have own property "CSSLab" expected property "CSSLab" missing
[FAIL] CSSLab interface: attribute alpha
  assert_own_property: self does not have own property "CSSLab" expected property "CSSLab" missing
[FAIL] CSSLCH interface: existence and properties of interface object
  assert_own_property: self does not have own property "CSSLCH" expected property "CSSLCH" missing
[FAIL] CSSLCH interface object length
  assert_own_property: self does not have own property "CSSLCH" expected property "CSSLCH" missing
[FAIL] CSSLCH interface object name
  assert_own_property: self does not have own property "CSSLCH" expected property "CSSLCH" missing
[FAIL] CSSLCH interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "CSSLCH" expected property "CSSLCH" missing
[FAIL] CSSLCH interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "CSSLCH" expected property "CSSLCH" missing
[FAIL] CSSLCH interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "CSSLCH" expected property "CSSLCH" missing
[FAIL] CSSLCH interface: attribute l
  assert_own_property: self does not have own property "CSSLCH" expected property "CSSLCH" missing
[FAIL] CSSLCH interface: attribute c
  assert_own_property: self does not have own property "CSSLCH" expected property "CSSLCH" missing
[FAIL] CSSLCH interface: attribute h
  assert_own_property: self does not have own property "CSSLCH" expected property "CSSLCH" missing
[FAIL] CSSLCH interface: attribute alpha
  assert_own_property: self does not have own property "CSSLCH" expected property "CSSLCH" missing
[FAIL] CSSOKLab interface: existence and properties of interface object
  assert_own_property: self does not have own property "CSSOKLab" expected property "CSSOKLab" missing
[FAIL] CSSOKLab interface object length
  assert_own_property: self does not have own property "CSSOKLab" expected property "CSSOKLab" missing
[FAIL] CSSOKLab interface object name
  assert_own_property: self does not have own property "CSSOKLab" expected property "CSSOKLab" missing
[FAIL] CSSOKLab interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "CSSOKLab" expected property "CSSOKLab" missing
[FAIL] CSSOKLab interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "CSSOKLab" expected property "CSSOKLab" missing
[FAIL] CSSOKLab interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "CSSOKLab" expected property "CSSOKLab" missing
[FAIL] CSSOKLab interface: attribute l
  assert_own_property: self does not have own property "CSSOKLab" expected property "CSSOKLab" missing
[FAIL] CSSOKLab interface: attribute a
  assert_own_property: self does not have own property "CSSOKLab" expected property "CSSOKLab" missing
[FAIL] CSSOKLab interface: attribute b
  assert_own_property: self does not have own property "CSSOKLab" expected property "CSSOKLab" missing
[FAIL] CSSOKLab interface: attribute alpha
  assert_own_property: self does not have own property "CSSOKLab" expected property "CSSOKLab" missing
[FAIL] CSSOKLCH interface: existence and properties of interface object
  assert_own_property: self does not have own property "CSSOKLCH" expected property "CSSOKLCH" missing
[FAIL] CSSOKLCH interface object length
  assert_own_property: self does not have own property "CSSOKLCH" expected property "CSSOKLCH" missing
[FAIL] CSSOKLCH interface object name
  assert_own_property: self does not have own property "CSSOKLCH" expected property "CSSOKLCH" missing
[FAIL] CSSOKLCH interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "CSSOKLCH" expected property "CSSOKLCH" missing
[FAIL] CSSOKLCH interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "CSSOKLCH" expected property "CSSOKLCH" missing
[FAIL] CSSOKLCH interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "CSSOKLCH" expected property "CSSOKLCH" missing
[FAIL] CSSOKLCH interface: attribute l
  assert_own_property: self does not have own property "CSSOKLCH" expected property "CSSOKLCH" missing
[FAIL] CSSOKLCH interface: attribute c
  assert_own_property: self does not have own property "CSSOKLCH" expected property "CSSOKLCH" missing
[FAIL] CSSOKLCH interface: attribute h
  assert_own_property: self does not have own property "CSSOKLCH" expected property "CSSOKLCH" missing
[FAIL] CSSOKLCH interface: attribute alpha
  assert_own_property: self does not have own property "CSSOKLCH" expected property "CSSOKLCH" missing
[FAIL] CSSColor interface: existence and properties of interface object
  assert_own_property: self does not have own property "CSSColor" expected property "CSSColor" missing
[FAIL] CSSColor interface object length
  assert_own_property: self does not have own property "CSSColor" expected property "CSSColor" missing
[FAIL] CSSColor interface object name
  assert_own_property: self does not have own property "CSSColor" expected property "CSSColor" missing
[FAIL] CSSColor interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "CSSColor" expected property "CSSColor" missing
[FAIL] CSSColor interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "CSSColor" expected property "CSSColor" missing
[FAIL] CSSColor interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "CSSColor" expected property "CSSColor" missing
[FAIL] CSSColor interface: attribute colorSpace
  assert_own_property: self does not have own property "CSSColor" expected property "CSSColor" missing
[FAIL] CSSColor interface: attribute channels
  assert_own_property: self does not have own property "CSSColor" expected property "CSSColor" missing
[FAIL] CSSColor interface: attribute alpha
  assert_own_property: self does not have own property "CSSColor" expected property "CSSColor" missing
Harness: the test ran to completion.
